src/auth/requirements.txt
aiohttp==3.8.3
aiohttp-jwt==0.6.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncpg==0.27.0
asyncpgsa==0.27.1
attrs==22.1.0
charset-normalizer==2.1.1
frozenlist==1.3.1
greenlet==2.0.1
idna==3.4
multidict==6.0.2
PyJWT==2.6.0
SQLAlchemy==1.4.43
yarl==1.8.1
